[Edit] 변수명 수정 및 수정한 구조체로 갈아끼우기 #35
